master: Avoid root owned files in user's home directory
The following command allows you to initialize a new Ansible role:

```
docker exec ansible-master ansible-galaxy init --init-path=/opt/ansible/roles <role>
```

But `/opt/ansible` is a bind mount to the user's home directory so the
new role files will be created as root.

This commit changes the master container's configuration to create an
`ansible` user using the same UID/GID as the user running the
environment and start the master container's process using this `ansible` user.